Backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
langchain-core
langchain-community
tavily-python
pydantic
python-dotenv
google-genai
pandas
